Fix make file errors and add conditional builds
New chart packages are now only built if there is a new version of in the
`Chart.yaml`.

Fix a few syntax errors in the build command of the Makefile.
Also, the make file now will only rebuild packages when they have a version new number in the `Chart.yaml`.

Signed-off-by: Aaron Spiegel <aaron.spiegel@emc.com>