[FLEX-436] Add DECKS one-touch deployment of SRS gateway resources
This change adds helm chart support for Docker registry secrets for the
SRS gateway CR's secondary deployments.

This docker registry secret support will also require a small, followup
change to the DECKS srsgateway controller to actually use the registry
secrets that this helm chart creates.

This change provides 2 ways of setting up Docker registry secrets for
the remote access pods and the SRS notifier pods:

(1) Re-use and existing Docker registry secret
    (i.e. "bring-your-own" registry secret).

    An example helm install setting for this is:
```
--set dockerSecret=my-existing-registry-secret
```

(2) Alternatively, a user can specify a dockerUsername and dockerPassword,
    and this Helm chart will automatically create a Docker registry
    secret.

    An example helm install command line setting to generate a Docker
    registry secret:
```
--set dockerUsername=janedoe --set dockerPassword=MyPassword
```

    The name of the registry secret that is generated will either be:
```
    <customResourceName>-docker-secret
```
if customResourceName is set; otherwise, it will be:
```
    <product>-docker-secret
```

If the Docker registries that are being used do not require
authentication, then the above helm chart settings can be left unset,
which is the default.