[FLEX-802] Upload helm and kubernetes resource configurations to SRS
This change is the Helm chart counterpart to this DECKS change:
https://asdstash.isus.emc.com/projects/VI/repos/decks/pull-requests/56

This change adds support for allowing the SRS Gateway CR controller
to automatically deploy a Kubernetes CronJob immediately after
successful registration with the SRS gateway. This "config-upload"
CronJob will periodically (default: every 12 hours) scrape Helm release
and Kubernetes resource config, and upload this config to the SRS
GW in the form of a compressed tar (.tar.gz) file.