[FLEX-1012] Fix Helm lint errors
This change fixes some helm lint errors for the SRS GW CR templates.

The root problem is that `helm lint` isn't smart enough to recognize that
when fields such as .Values.product are set as required fields, they can
never be undefined. Instead, `helm lint` does its analysis treating
these required fields as potentially undefined.

The fix is to add an (effectively unnecessary in this case) conditional test
for the required fields being undefined before the fields are used, and
adding a conditional branch for the fields not being defined that can
never be reached.