[FLEX-1156] added deos_events as a part of ecs_cluster chart (#106)
This PR adds does-events as a part of ECSCluster. The deos-events container is used to pull the events from the object store and convert them into k8s events. This pod will run after the provisioning job is done.The operator side of the code changes will have separate PR